Dots now collide VERY well
It turns out you don't need rigid body physics to do stuff.

Basically, I changed the forces caused by other dots to only come into
effect when the distance between two dots is around the sum of their
radii. Then, calculate the angle of the line between the two dots,
set a force magnitude, and calculate the x and y components of the force
to add to forces.

I also quickly set it up to randomize the targetPositions for the dots
to go to every 100 ticks and made color dependent on what
targetPosition the dot is aiming at.

Overall, this works WAAAAAYYYY better than it deserves to