added a quick implementation of rigid bodies.
it's a little cooked and looks like it's boiling if there's too many
dots crowding around the same place, but most of the time it looks fine.

I wanna add better interactions in terms of transferring energy/momentum
between dots when the collide, which might solve some of the boiling
issues and make it look more like a fluid.